+++ frontend/src/components/Dropdown.tsx
@@ -1,6 +1,7 @@
-import React, { useState } from "react";
+import React from "react";
 import styled from "@emotion/styled";
 import { keyframes } from "@emotion/react";
+import useDropdown from "@/hooks/useDropdown";
 
 interface optionList {
   option: string;
@@ -12,16 +13,7 @@ interface DropdownProp {
 }
 
 const Dropdown = ({ DropdownButton, optionList }: DropdownProp) => {
-  const [isOpened, setIsOpened] = useState(false);
-
-  const handleButtonClick = () => {
-    setIsOpened((prev) => !prev);
-  };
-
-  const handleOptionClick = (callback: () => void) => {
-    setIsOpened(false);
-    callback();
-  };
+  const { isOpened, handleButtonClick, handleOptionClick } = useDropdown();
 
   return (
     <StyledDropdown>
@@ -54,7 +46,6 @@ const fadeInDown = keyframes`
     opacity: 0;
     margin-top: -10px;
   }
-
   to {
     opacity: 1;
     transform: translate3d(0, 0, 0);
@@ -66,17 +57,12 @@ const StyledDropdownList = styled.ul`
   top: 24px;
   width: 120px;
   padding: 8px;
-
   z-index: 2;
-
   text-align: center;
-
   background-color: ${({ theme }) => theme.colors.WHITE};
   box-shadow: 0 4px 4px rgba(89, 87, 87, 0.25);
   border-radius: 8px;
-
   animation: ${fadeInDown} 0.3s;
-
   hr {
     border: 0.5px solid ${({ theme }) => theme.colors.GRAY_400};
   }
@@ -84,7 +70,6 @@ const StyledDropdownList = styled.ul`
 
 const StyledOption = styled.li`
   font-size: 14px;
-
   cursor: pointer;
 `;
